refactor(cosh-ng): split i18n message ids
