+++ src/cosh-ng/crates/cosh-shell/src/i18n/en/config.rs
@@ -1,0 +1,33 @@
+use super::MessageId;
+
+pub(super) fn message(id: MessageId) -> Option<&'static str> {
+    Some(match id {
+        MessageId::ConfigInvalidLanguageBody => "Invalid language: {language}",
+        MessageId::ConfigSupportedLanguagesFooter => "Supported: auto, en-US, zh-CN.",
+        MessageId::ConfigUnknownKeyBody => "Unknown config key: {key}",
+        MessageId::ConfigHomeMissingBody => "HOME is not set; cannot persist config.",
+        MessageId::ConfigHomeMissingFooter => "Set HOME or edit config manually.",
+        MessageId::ConfigUnchangedTitle => "Config unchanged",
+        MessageId::ConfigNoFileChangedBody => "No config file was changed.",
+        MessageId::ConfigSavedTitle => "Config saved",
+        MessageId::ConfigSavedValueLine => "Saved ui.{setting} = \"{value}\".",
+        MessageId::ConfigCurrentSessionLanguageLine => "Current session language: {language}.",
+        MessageId::ConfigSavedFooter => "Saved setting takes effect next startup.",
+        MessageId::ConfigSaveFailedTitle => "Config save failed",
+        MessageId::ConfigSaveFailedBody => "Config save failed: {error}",
+        MessageId::ConfigSavePromptTitle => "Save config?",
+        MessageId::ConfigFileLine => "file: {path}",
+        MessageId::ConfigPendingChangeLine => "ui.{setting}: {before} -> {after}",
+        MessageId::ConfigSaveButton => "Save",
+        MessageId::ConfigCancelButton => "Cancel",
+        MessageId::ConfigApplyKeysFooter => "Keys: Left/Right select | Enter apply | Esc cancel",
+        MessageId::ConfigLanguageTitle => "Language",
+        MessageId::ConfigLanguageAutoLine => "auto    Follow LC_ALL/LC_MESSAGES/LANG",
+        MessageId::ConfigLanguageEnLine => "en-US   English",
+        MessageId::ConfigLanguageZhLine => "zh-CN   Simplified Chinese",
+        MessageId::ConfigLanguageKeysFooter => {
+            "Keys: Left/Right select | Enter choose | Esc cancel"
+        }
+        _ => return None,
+    })
+}
