+++ src/cosh-ng/crates/cosh-shell/src/i18n/message_id.rs
@@ -1,700 +1,79 @@
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum MessageId {
-    StartupTitle,
-    StartupAdapterLine,
-    StartupCwdLine,
-    StartupCommandsLine,
-    StartupHooksNoneSummary,
-    StartupHooksCompletedSummary,
-    StartupHooksFindingsHeading,
-    StartupHooksRustProjectFinding,
-    StartupHooksNoFindings,
-    StartupHooksReadOnlyNote,
-    StartupSwitchHint,
-    HelpTitle,
-    HelpFooter,
-    HelpGroupConfig,
-    HelpGroupModes,
-    HelpGroupHooks,
-    HelpSummaryHelp,
-    HelpSummaryAuth,
-    HelpSummaryConfig,
-    HelpSummaryRecommendations,
-    HelpSummaryModeApproval,
-    HelpSummaryModeAnalysis,
-    HelpSummaryAgent,
-    HelpSummaryExplain,
-    HelpSummaryCancel,
-    HelpSummaryDetails,
-    HelpSummaryAudit,
-    HelpSummaryHooks,
-    HelpSummarySelect,
-    HelpSummaryCopy,
-    HelpSummaryDebug,
-    HelpSummaryClear,
-    HelpSummaryShell,
-    HelpSummaryApprovalModeRemoved,
-    SlashHintTitle,
-    SlashHintPrefix,
-    SlashHintCurrentMode,
-    SlashHintFooter,
-    SlashUnknownTitle,
-    SlashUnknownBody,
-    SlashUnknownSuggestionBody,
-    SlashUnknownFooter,
-    SlashInfoAuditTitle,
-    SlashInfoAuditApprovalsBody,
-    SlashInfoAuditActivityBody,
-    SlashInfoAuditFooter,
-    SlashInfoConfigTitle,
-    SlashInfoConfigLanguageLine,
-    SlashInfoConfigLanguageEffectiveLine,
-    SlashInfoConfigPathLine,
-    SlashInfoConfigDebugActivityLine,
-    SlashInfoConfigAnalysisStrategyLine,
-    SlashInfoConfigRenderFallbackLine,
-    SlashInfoConfigFooter,
-    ConfigInvalidLanguageBody,
-    ConfigSupportedLanguagesFooter,
-    ConfigUnknownKeyBody,
-    ConfigHomeMissingBody,
-    ConfigHomeMissingFooter,
-    ConfigUnchangedTitle,
-    ConfigNoFileChangedBody,
-    ConfigSavedTitle,
-    ConfigSavedValueLine,
-    ConfigCurrentSessionLanguageLine,
-    ConfigSavedFooter,
-    ConfigSaveFailedTitle,
-    ConfigSaveFailedBody,
-    ConfigSavePromptTitle,
-    ConfigFileLine,
-    ConfigPendingChangeLine,
-    ConfigSaveButton,
-    ConfigCancelButton,
-    ConfigApplyKeysFooter,
-    ConfigLanguageTitle,
-    ConfigLanguageAutoLine,
-    ConfigLanguageEnLine,
-    ConfigLanguageZhLine,
-    ConfigLanguageKeysFooter,
-    SlashHooksRegisteredTitle,
-    SlashHooksNoHooksBody,
-    SlashHooksStatusCountLine,
-    SlashHooksStatusSourcesLine,
-    SlashHooksStatusProjectTrustLine,
-    SlashHooksFooterCount,
-    SlashHooksFooterMutedTargets,
-    SlashHooksTargetMutedTitle,
-    SlashHooksTargetMutedBody,
-    SlashHooksTargetMutedFooter,
-    SlashHooksTargetUnmutedTitle,
-    SlashHooksTargetUnmutedBody,
-    SlashHooksTargetNotMutedBody,
-    SlashHooksEnabledTitle,
-    SlashHooksEnabledBody,
-    SlashHooksDisabledTitle,
-    SlashHooksDisabledBody,
-    SlashHooksHistoryTitle,
-    SlashHooksHistoryEmptyBody,
-    SlashHooksHistoryFooter,
-    SlashHooksEventsTitle,
-    SlashHooksEventsEmptyBody,
-    SlashHooksEventsFooter,
-    SlashHooksUsageTitle,
-    SlashHooksUsageListLine,
-    SlashHooksUsageHistoryLine,
-    SlashHooksUsageEventsLine,
-    SlashHooksUsageAnalyzeLine,
-    SlashHooksUsageIgnoreLine,
-    SlashHooksUsageDetailsLine,
-    SlashHooksUsageFeedbackLine,
-    SlashHooksUsageClearFeedbackLine,
-    SlashHooksUsageMuteLine,
-    SlashHooksUsageUnmuteLine,
-    SlashHooksUsageTrustProjectLine,
-    SlashHooksUsageUntrustProjectLine,
-    SlashHooksUsageClearProjectTrustLine,
-    SlashHooksUsageEnableLine,
-    SlashHooksUsageDisableLine,
-    SlashHooksProjectTrustedTitle,
-    SlashHooksProjectUntrustedTitle,
-    SlashHooksProjectTrustNoHooksBody,
-    SlashHooksProjectTrustedBody,
-    SlashHooksProjectUntrustedBody,
-    SlashHooksProjectTrustNoChangeFooter,
-    SlashHooksProjectTrustPersistedFooter,
-    SlashHooksProjectTrustRemovedFooter,
-    SlashHooksProjectTrustPersistenceFailedFooter,
-    SlashHooksProjectTrustClearedTitle,
-    SlashHooksProjectTrustClearedBody,
-    SlashHooksProjectTrustClearedFooter,
-    SlashHooksProjectTrustClearFailedFooter,
-    SlashHooksFeedbackUsageBody,
-    SlashHooksFeedbackTitle,
-    SlashHooksFeedbackFindingNotFoundBody,
-    SlashHooksFeedbackFindingNotFoundFooter,
-    SlashHooksFeedbackRecordedTitle,
-    SlashHooksFeedbackRecordedBody,
-    SlashHooksFeedbackHookLine,
-    SlashHooksFeedbackPolicyKeyLine,
-    SlashHooksFeedbackPersistedFooter,
-    SlashHooksFeedbackPersistenceFailedFooter,
-    SlashHooksFeedbackClearedTitle,
-    SlashHooksFeedbackClearedBody,
-    SlashHooksFeedbackClearedFooter,
-    SlashHooksFeedbackClearFailedFooter,
-    DebugSessionTitle,
-    DebugAdapterLine,
-    DebugProviderInvocationLine,
-    DebugProviderCommittedSessionLine,
-    DebugActiveRunLine,
-    DebugQueuedRunsLine,
-    DebugProviderPendingSessionLine,
-    DebugProviderInitializeSeenLine,
-    DebugHostExecutedShellResultLine,
-    DebugSelectedShellExecutionPathLine,
-    DebugLatestProviderRequestLine,
-    DebugLatestToolUseLine,
-    DebugLatestRecoveryStatusLine,
-    DebugLatestRecoveryReasonLine,
-    DebugEvidenceAccessLine,
-    DebugEvidenceToolRegisteredLine,
-    DebugEvidenceNamespaceLine,
-    DebugEvidenceLedgerCountLine,
-    DebugLatestShellOutputReadLine,
-    DebugUnknownTargetBody,
-    DebugUnknownTargetFooter,
-    CommandRemovedTitle,
-    ApprovalModeRemovedBody,
-    ApprovalModeRemovedFooter,
-    RemovedDecisionCommandBody,
-    RemovedApprovalDecisionFooter,
-    RemovedQuestionAnswerFooter,
-    ModeTitle,
-    ModesTitle,
-    ModeApprovalLine,
-    ModeAnalysisLine,
-    ModeSummaryFooter,
-    ModeRemovedTitle,
-    ModeRemovedBody,
-    ModeRemovedFooter,
-    ModeLanguageBody,
-    ModeLanguageFooter,
-    ModeUnknownBody,
-    ModeUnknownFooter,
-    ApprovalModeTitle,
-    ApprovalModeSetBody,
-    ApprovalModeUnknownBody,
-    ApprovalModeUsageFooter,
-    ApprovalModeRecommendFooter,
-    ApprovalModeAutoFooter,
-    ApprovalModeTrustFooter,
-    ApprovalModeTrustConfirmationTitle,
-    ApprovalModeTrustConfirmationBody,
-    ApprovalModeTrustConfirmationCommandBody,
-    ApprovalModeTrustConfirmationFooter,
-    ApprovalModeCardTitle,
-    ApprovalModeCardCurrentLine,
-    ApprovalModeCardRecommendLine,
-    ApprovalModeCardAutoLine,
-    ApprovalModeCardTrustLine,
-    ApprovalModeCardFooter,
-    ApprovalModeRemainsBody,
-    ApprovalModeCancelBody,
-    ApprovalModeCancelFooter,
-    AnalysisModeTitle,
-    AnalysisModeCurrentBody,
-    AnalysisModeSetBody,
-    AnalysisModeUnknownBody,
-    AnalysisModeUsageFooter,
-    AnalysisModeSmartFooter,
-    AnalysisModeAutoFooter,
-    AnalysisModeManualFooter,
-    AnalysisModeCardSmartLine,
-    AnalysisModeCardAutoLine,
-    AnalysisModeCardManualLine,
-    AnalysisModeCardFooter,
-    AnalysisModeRemainsBody,
-    AnalysisModeCancelBody,
-    AnalysisModeCancelFooter,
-    AgentThinking,
-    AgentThinkingElapsed,
-    AgentRecoveryTitle,
-    AgentRecoveryFreshTurnBody,
-    AgentRecoveryContinuityBody,
-    AgentStatusTitle,
-    AgentStillWorking,
-    AgentStatusFooter,
-    AgentStatusStarting,
-    AgentStatusWaitingBackend,
-    AgentStatusThinking,
-    AgentStatusPreparingModelSession,
-    AgentStatusStartingModelBackend,
-    AgentStatusModelInitialized,
-    AgentStatusModelStatus,
-    AgentStatusAnalysisCompleted,
-    AgentStatusAnalysisReturnedError,
-    AgentStatusStreaming,
-    AgentStatusReceivingResponse,
-    AgentStatusApproval,
-    AgentStatusWaitingApprovalTool,
-    AgentStatusQuestion,
-    AgentStatusWaitingUserAnswer,
-    AgentStatusWaitingApprovalCommand,
-    AgentStatusTool,
-    AgentStatusCapturingToolOutput,
-    AgentStatusToolCompleted,
-    AgentStatusCompleted,
-    AgentStatusFailed,
-    AgentStatusCancelled,
-    AgentStatusRunningApprovedProviderTool,
-    AgentProviderTimeoutDroppedQueuedBody,
-    AgentCancellationRequestedTitle,
-    AgentCancellationRequestedBody,
-    AgentCancelledReasonLabel,
-    AgentCancelledUserRequestedReason,
-    AgentResponseTitle,
-    AgentGovernanceTitle,
-    AgentGovernanceStatusLine,
-    AgentGovernanceReasonLine,
-    AgentGovernanceSummaryLine,
-    AgentGovernanceErrorLine,
-    AgentGovernanceToolOutputLine,
-    AgentGovernanceToolCompletedLine,
-    AgentGovernanceApprovalRequiredLine,
-    AgentGovernanceShellCommandSubject,
-    AgentGovernanceBashCommandSubject,
-    AgentGovernanceToolSubject,
-    AgentGovernanceBlockedUserApprovalLine,
-    AgentGovernanceQuestionLine,
-    AgentRecommendedCommandsLabel,
-    InterceptNoticeTitle,
-    InterceptNoticeBody,
-    InterceptNoticeFooter,
-    FailedAnalysisCancelledTitle,
-    FailedAnalysisCancelledBody,
-    FailedAnalysisCancelNoActiveBody,
-    FailedAnalysisCancelledFooter,
-    AnalysisSkippedTitle,
-    AnalysisSkippedBody,
-    AnalysisSkippedFooter,
-    HookAutoAnalyzedTitle,
-    HookAutoAnalyzedBody,
-    HookAutoAnalyzedFooter,
-    InsightLabel,
-    InsightCommandTypoSummary,
-    InsightPermissionDeniedSummary,
-    InsightBuildOrTestFailureSummary,
-    InsightRuntimeExceptionSummary,
-    InsightAbnormalSignalSummary,
-    InsightMemoryPressureSummary,
-    InsightHighMemoryProcessSummary,
-    InsightHighMemoryProcessGenericSummary,
-    InsightMemoryRootCauseSummary,
-    InsightMemoryRootCauseGenericSummary,
-    InsightPermissionDeniedPrompt,
-    InsightBuildOrTestFailurePrompt,
-    InsightRuntimeExceptionPrompt,
-    InsightAbnormalSignalPrompt,
-    InsightMemoryPressurePrompt,
-    InsightHighMemoryProcessPrompt,
-    InsightHighMemoryProcessGenericPrompt,
-    InsightMemoryRootCausePrompt,
-    InsightMemoryRootCauseGenericPrompt,
-    InsightShellRewriteFirstUseHint,
-    InsightAgentPromptFirstUseHint,
-    AgentQueuedTitle,
-    AgentQueuedBodyCommand,
-    AgentQueuedBodyActive,
-    AgentQueuedFooter,
-    HookFindingTitle,
-    HookFindingFooter,
-    HookFindingMarkdownTitle,
-    HookFindingMarkdownHookLine,
-    HookFindingMarkdownSeverityLine,
-    HookFindingMarkdownFindingLine,
-    HookFindingMarkdownSuggestionLine,
-    HookFindingMarkdownRelatedTitle,
-    HookFindingMarkdownRelatedLine,
-    HookFindingMarkdownAgentFollowUpLine,
-    HookHintTitle,
-    HookHintNotFoundBody,
-    HookHintNotFoundFooter,
-    HookHintNoFindingBody,
-    HookHintBlockUnavailableBody,
-    HookHintIgnoredTitle,
-    HookHintIgnoredBody,
-    HookHintIgnoredFooter,
-    HookHintUsageTitle,
-    HookHintUsageBody,
-    HookFindingDetailsTitle,
-    HookConsultationHookLabel,
-    HookConsultationConfidenceReasonLine,
-    HookConsultationFindingLine,
-    HookConsultationSuggestionLine,
-    HookConsultationAnalyzeAction,
-    HookConsultationIgnoreAction,
-    HookDetailsConfidenceLine,
-    HookDetailsUserInterestLine,
-    HookDetailsReasonLookupIntent,
-    HookDetailsReasonPipelineIntent,
-    HookDetailsReasonScriptIntent,
-    HookDetailsReasonWrapperLowConfidence,
-    HookDetailsReasonInteractiveIntent,
-    HookDetailsReasonActiveRunDeferred,
-    HookDetailsReasonUserContinuedInput,
-    HookDetailsReasonNonDiagnosticSuccessCommand,
-    HookDetailsReasonFeedbackNoisy,
-    HookDetailsReasonIgnoredSameFinding,
-    HookDetailsReasonSameCardAlreadyRendered,
-    HookDetailsReasonInterruptionBudget,
-    HookDetailsReasonLowConfidence,
-    HookDetailsReasonDiagnosticIntent,
-    HookDetailsReasonOtherIntent,
-    HookDetailsTopicLine,
-    HookDetailsOriginLine,
-    HookDetailsSuppressionKeyLine,
-    HookDetailsOutputRefLine,
-    HookDetailsCreatedAtLine,
-    HookDetailsPromptHintLine,
-    HookDetailsRecommendedSkillLine,
-    HookDetailsReadOnlyCliHintLine,
-    HookDetailsFooter,
-    RuntimeDetailsUnavailableTitle,
-    RuntimeDetailsUnavailableBody,
-    ActivityTitle,
-    ActivityDetailsTitle,
-    ActivityRunLabel,
-    ActivityDetailLabel,
-    ActivitySkillLabel,
-    ActivitySkillUpdatedStatus,
-    ActivityToolLabel,
-    ActivityToolOutputLabel,
-    ActivityShellLabel,
-    ActivityStatusLoading,
-    ActivityStatusLoaded,
-    ActivityStatusFailed,
-    ActivityStatusCalled,
-    ActivityStatusRequested,
-    ActivityStatusCaptured,
-    ActivityStatusCompleted,
-    ActivityStatusError,
-    ActivityStatusInterrupted,
-    ActivityToolCalledSummary,
-    ActivityToolRequestedSummary,
-    ActivityToolOutputCapturedSummary,
-    ActivityProviderNativeShellBypassSummary,
-    ActivityToolNeedsForegroundShellSummary,
-    ActivityShellHandoffSentSummary,
-    ToolCardReadFileLabel,
-    ToolCardWriteFileLabel,
-    ToolCardEditFileLabel,
-    ToolCardSearchFilesLabel,
-    ToolCardFindFilesLabel,
-    ToolCardListDirectoryLabel,
-    ToolCardShellLabel,
-    ToolCardWebFetchLabel,
-    ToolCardWebSearchLabel,
-    ToolCardSkillLabel,
-    ToolCardAgentLabel,
-    ToolCardMemoryLabel,
-    ToolCardEvidenceLabel,
-    ToolCardCustomToolLabel,
-    ToolCardCalledStatus,
-    ToolCardRequestedStatus,
-    ToolCardAutoApprovedStatus,
-    ToolCardCapturedStatus,
-    ToolCardCompletedStatus,
-    ToolCardFailedStatus,
-    ToolCardDuplicateStatus,
-    ToolCardInterruptedStatus,
-    ToolCardReadOnlyIntent,
-    ToolCardWriteIntent,
-    ToolCardExecuteIntent,
-    ToolCardNetworkIntent,
-    ToolCardContextIntent,
-    ToolCardCustomIntent,
-    ToolCardApprovalRequiredAction,
-    ToolCardWriteCompletedResult,
-    ToolCardEditCompletedResult,
-    ToolCardSkillAvailableResult,
-    ToolCardShellEvidenceDeliveredResult,
-    ToolCardShellEvidenceListResult,
-    ToolCardShellEvidenceReadResult,
-    ToolCardShellEvidenceAlreadyDeliveredResult,
-    ToolCardShellEvidenceFailedResult,
-    ToolCardShellEvidenceDuplicateResult,
-    ToolCardShellEvidenceMetadataMetric,
-    ToolCardOutputCapturedResult,
-    ToolCardLinesReturnedResult,
-    ToolCardStdoutMetric,
-    ToolCardStderrMetric,
-    ToolCardTruncatedMetric,
-    MarkdownCodeLabel,
-    MarkdownCodeWithLanguageLabel,
-    MarkdownTableLabel,
-    RecommendationTitle,
-    RecommendationNextStepTitle,
-    AnalysisResultTitle,
-    RecommendationEmptyBody,
-    RecommendationFooter,
-    RecommendationNoSelectableTitle,
-    RecommendationNoSelectableBody,
-    RecommendationUnavailableTitle,
-    RecommendationUnavailableBody,
-    RecommendationSelectedTitle,
-    RecommendationSelectedBody,
-    RecommendationCopiedTitle,
-    RecommendationCopiedBody,
-    RecommendationInsertTitle,
-    RecommendationInsertBody,
-    RecommendationDetailsTitle,
-    RecommendationDetailsBody,
-    RecommendationDisplayOnlyBody,
-    RecommendationCopyOnlyBody,
-    RecommendationInsertOnlyBody,
-    RecommendationDetailsOnlyBody,
-    HealthBannerTitle,
-    HealthStartupLabel,
-    HealthBannerTryLabel,
-    HealthBannerFindingLabel,
-    HealthBannerEvidenceLabel,
-    HealthBannerMoreFindingsLabel,
-    HealthBannerUnavailableLabel,
-    HealthBannerFindingsSection,
-    HealthBannerSuggestedPromptSection,
-    HealthBannerSuggestedPromptIntro,
-    HealthSeverityOk,
-    HealthSeverityWarning,
-    HealthSeverityCritical,
-    HealthSeverityDegraded,
-    HealthSeverityUnavailable,
-    HealthMetricCpu,
-    HealthMetricCpuLoadPerCore,
-    HealthMetricLoad1mShort,
-    HealthMetricLoad5mShort,
-    HealthMetricLoadValue,
-    HealthMetricCpuPerCoreUnit,
-    HealthMetricCpuUsed,
-    HealthMetricHost,
-    HealthMetricMemory,
-    HealthMetricMemoryAvailable,
-    HealthMetricMemoryUsed,
-    HealthMetricSwap,
-    HealthMetricSwapUsed,
-    HealthMetricDisk,
-    HealthMetricDiskUsed,
-    HealthMetricDiskMountUsed,
-    HealthMetricPressure,
-    HealthMetricLevels,
-    HealthMetricSignal,
-    HealthMetricService,
-    HealthEvidenceDiskAvailable,
-    HealthEvidenceMount,
-    HealthEvidenceOomKilledProcess,
-    HealthEvidenceOomCgroup,
-    HealthEvidenceOomOneHourCount,
-    HealthEvidenceOomTwentyFourHourCount,
-    HealthEvidenceOomVictimKilledAgo,
-    HealthEvidenceOomVictimKilled,
-    HealthEvidenceOomAge,
-    HealthOomScopeMemcg,
-    HealthOomScopeHost,
-    HealthOomScopeCpuset,
-    HealthOomScopeMemoryPolicy,
-    HealthOomScopeUnknown,
-    HealthTryAnalyzeMemoryPressure,
-    HealthTryCheckSwapPressure,
-    HealthTryCheckRecentOom,
-    HealthTryInspectDiskUsage,
-    HealthTryInspectServiceStatus,
-    HealthTryInspectHighLoad,
-    HealthTryInspectProcessMemory,
-    HealthTryReviewUnavailableChecks,
-    HealthPromptAnalyzeMemoryPressure,
-    HealthPromptCheckSwapPressure,
-    HealthPromptCheckRecentOom,
-    HealthPromptInspectDiskUsage,
-    HealthPromptInspectServiceStatus,
-    HealthPromptInspectHighLoad,
-    HealthPromptInspectProcessMemory,
-    HealthPromptReviewUnavailableChecks,
-    HealthInsightMemoryAvailableLow,
-    HealthInsightDiskHigh,
-    HealthInsightRecentOom,
-    HealthInsightCpuLoadHigh,
-    HealthInsightSwapPressure,
-    HealthInsightServiceState,
-    HealthInsightGeneric,
-    HealthUnavailablePermissionDenied,
-    HealthUnavailableCommandMissing,
-    HealthUnavailableTimeout,
-    HealthUnavailableUnsupported,
-    HealthUnavailableParseError,
-    HealthFindingPlatformUnsupported,
-    HealthFindingCoreCollectorUnavailable,
-    HealthFindingCpuLoadHigh,
-    HealthFindingMemoryAvailableLow,
-    HealthFindingSwapPressure,
-    HealthFindingDiskHigh,
-    HealthFindingRecentOom,
-    HealthFindingKernelPanic,
-    HealthFindingServiceFailed,
-    HealthFindingServiceInactive,
-    HealthTryReasonMemoryLow,
-    HealthTryReasonSwapWithContext,
-    HealthTryReasonRecentOom,
-    HealthTryReasonDiskHigh,
-    HealthTryReasonServiceState,
-    HealthTryReasonHighLoad,
-    HealthTryReasonMissingCoreCheck,
-    ToolOutputStdoutCapturedSummary,
-    ToolOutputStderrCapturedSummary,
-    ToolSummaryExit,
-    ToolSummaryBlocked,
-    ToolSummaryTimedOut,
-    ToolSummaryFailed,
-    QuestionTitle,
-    QuestionDefaultPrompt,
-    QuestionAnswerLabel,
-    QuestionSelectOneLabel,
-    QuestionSelectMultipleLabel,
-    QuestionOtherEmptyLabel,
-    QuestionKeysPrefix,
-    QuestionInstructionMoveTypeSend,
-    QuestionInstructionMoveToggleSend,
-    QuestionInstructionMoveSend,
-    QuestionInstructionTypeSend,
-    QuestionInstructionNoAnswer,
-    QuestionNoPendingTitle,
-    QuestionNoPendingBody,
-    ApprovalTitle,
-    ApprovalRequiredTitle,
-    ApprovalResolutionApprovedTitle,
-    ApprovalResolutionAutoApprovedTitle,
-    ApprovalResolutionTrustedTitle,
-    ApprovalResolutionDeniedTitle,
-    ApprovalResolutionCancelledTitle,
-    ApprovalResolutionBlockedTitle,
-    ApprovalResolutionDeferredTitle,
-    ApprovalActionAllowOnce,
-    ApprovalActionAlwaysTrust,
-    ApprovalActionDeny,
-    ApprovalActionDetails,
-    ApprovalToolInputLabel,
-    ApprovalCommandLabel,
-    ApprovalDetailsTitle,
-    ApprovalDetailsSourceLabel,
-    ApprovalDetailsRunLabel,
-    ApprovalDetailsExecutionLabel,
-    ApprovalDetailsCommandBlockLabel,
-    ApprovalDetailsRedactionLabel,
-    ApprovalDetailsProviderRequestLabel,
-    ApprovalDetailsToolUseLabel,
-    ApprovalDetailsDefaultDenyLine,
-    ApprovalDetailsRequestLabel,
-    ApprovalDetailsInputLabel,
-    ApprovalDetailsBashCommandSubject,
-    ApprovalDetailsShellCommandSubject,
-    ApprovalDetailsToolSubject,
-    ApprovalDetailsPendingValue,
-    ApprovalDetailsNoneValue,
-    ApprovalDetailsNotApplicableValue,
-    ApprovalAssessmentSummaryLine,
-    ApprovalAssessmentReasonLine,
-    ApprovalJournalTitle,
-    ApprovalJournalDecisionCount,
-    ApprovalJournalEmptyBody,
-    ApprovalJournalActorLabel,
-    ApprovalJournalPreviewHashLabel,
-    ApprovalJournalSubjectLabel,
-    ApprovalJournalPreviewLabel,
-    ApprovalRiskSuffix,
-    ApprovalQueueCompactLine,
-    ApprovalQueueFullLine,
-    ApprovalQueueNextSuffix,
-    ApprovalSubjectLabel,
-    ApprovalNextLabel,
-    ApprovalKeysPrefix,
-    ApprovalKeysText,
-    ApprovalExecutableToolPolicy,
-    ApprovalExecutableToolPolicyExtra,
-    ApprovalCommandDefaultPolicy,
-    ApprovalRunShellCommandPrompt,
-    ApprovalRunBashCommandPrompt,
-    ApprovalNotFoundTitle,
-    ApprovalNotFoundBody,
-    ApprovalShellHandoffNotFoundTitle,
-    ApprovalShellHandoffNotFoundBody,
-    ApprovalShellHandoffBlockedTitle,
-    ApprovalShellHandoffBlockedFooter,
-    ApprovalShellHandoffValidationEmptyCommand,
-    ApprovalShellHandoffValidationMultilineCommand,
-    ApprovalShellHandoffValidationControlCharacter,
-    ApprovalShellHandoffValidationEmptyPreview,
-    ApprovalShellHandoffValidationEmptyApprovalId,
-    ApprovalShellHandoffValidationEmptyRunId,
-    ApprovalShellHandoffSendingTitle,
-    ApprovalShellHandoffSendingBody,
-    ApprovalShellHandoffTimeoutTitle,
-    ApprovalShellHandoffTimeoutExceededBody,
-    ApprovalShellHandoffTimeoutInterruptBody,
-    ApprovalReceiptKindToolRequest,
-    ApprovalReceiptKindShellCommandRequest,
-    ApprovalReceiptKindBashTool,
-    ApprovalReceiptDecisionPending,
-    ApprovalReceiptDecisionApproved,
-    ApprovalReceiptDecisionSentToShell,
-    ApprovalReceiptDecisionProviderNativeAllowed,
-    ApprovalReceiptDecisionApprovedDisplayOnly,
-    ApprovalReceiptDecisionDenied,
-    ApprovalReceiptDecisionCancelled,
-    ApprovalReceiptDecisionBlocked,
-    ApprovalReceiptSubjectBashSentToShell,
-    ApprovalReceiptSubjectBashProviderNative,
-    ApprovalReceiptBashSentToShellMessage,
-    ApprovalReceiptProviderNativeAllowedMessage,
-    ApprovalHookHeading,
-    HelpGroupSessions,
-    HelpSummarySession,
-    SessionTitle,
-    SessionUnavailableBody,
-    SessionBusyBody,
-    SessionEmptyBody,
-    SessionListFooter,
-    SessionStatusTitle,
-    SessionShellIdLine,
-    SessionProviderIdLine,
-    SessionWorkspaceLine,
-    SessionRecoveryLine,
-    SessionErrorLine,
-    SessionEvidenceNotRestoredBody,
-    SessionPickerFooter,
-    SessionClearConfirmTitle,
-    SessionClearConfirmCountLine,
-    SessionClearConfirmFooter,
-    SessionSelectedTitle,
-    SessionSelectedBody,
-    SessionErrorTitle,
-    SessionClearedTitle,
-    SessionClearedBody,
-    SessionSkippedBody,
-    SessionClearInterruptedBody,
-    SessionCancelledTitle,
-    SessionCancelledBody,
-    SessionUsageBody,
-    SessionNotReadyBody,
-    SessionProtectedBody,
-    HelpGroupRegistry,
-    HelpSummaryExtensions,
-    HelpSummarySkills,
-    SlashExtensionsTitle,
-    SlashSkillsTitle,
-    SlashRegistryUnavailable,
-    SlashHooksShellSection,
-    SlashHooksAgentSection,
-    SlashHooksAgentUnavailable,
-    SlashExtensionsEmptyBody,
-    SlashSkillsEmptyBody,
+#[macro_use]
+mod startup;
+#[macro_use]
+mod help;
+#[macro_use]
+mod config;
+#[macro_use]
+mod hooks;
+#[macro_use]
+mod debug;
+#[macro_use]
+mod modes;
+#[macro_use]
+mod agent;
+#[macro_use]
+mod insight;
+#[macro_use]
+mod hook_details;
+#[macro_use]
+mod activity;
+#[macro_use]
+mod recommendation;
+#[macro_use]
+mod health;
+#[macro_use]
+mod question;
+#[macro_use]
+mod approval;
+#[macro_use]
+mod session;
+
+macro_rules! define_message_id {
+    ($($id:ident,)*) => {
+        #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+        pub enum MessageId {
+            $($id,)*
+        }
+
+        impl MessageId {
+            pub const ALL: &'static [MessageId] = &[
+                $(MessageId::$id,)*
+            ];
+        }
+    };
 }
+
+macro_rules! collect_message_ids {
+    ([$segment:ident $(, $remaining:ident)* $(,)?], $($ids:ident,)*) => {
+        $segment!(collect_message_ids, [$($remaining),*], $($ids,)*);
+    };
+    ([], $($ids:ident,)*) => {
+        define_message_id!($($ids,)*);
+    };
+}
+
+// Segment order preserves the public fieldless enum's existing discriminants.
+collect_message_ids!([
+    startup_ids,
+    help_core_ids,
+    config_ids,
+    hooks_command_ids,
+    debug_ids,
+    legacy_approval_mode_ids,
+    removed_command_ids,
+    mode_ids,
+    agent_ids,
+    hook_insight_ids,
+    agent_queue_ids,
+    hook_details_ids,
+    activity_ids,
+    recommendation_ids,
+    health_ids,
+    tool_summary_ids,
+    question_ids,
+    approval_ids,
+    help_session_ids,
+    session_ids,
+    help_registry_ids,
+],);
