+++ src/cosh-ng/crates/cosh-shell/src/i18n/message_id/approval.rs
@@ -1,0 +1,95 @@
+macro_rules! approval_ids {
+    ($next:ident, $remaining:tt, $($ids:ident,)*) => {
+        $next!(
+            $remaining,
+            $($ids,)*
+            ApprovalTitle,
+            ApprovalRequiredTitle,
+            ApprovalResolutionApprovedTitle,
+            ApprovalResolutionAutoApprovedTitle,
+            ApprovalResolutionTrustedTitle,
+            ApprovalResolutionDeniedTitle,
+            ApprovalResolutionCancelledTitle,
+            ApprovalResolutionBlockedTitle,
+            ApprovalResolutionDeferredTitle,
+            ApprovalActionAllowOnce,
+            ApprovalActionAlwaysTrust,
+            ApprovalActionDeny,
+            ApprovalActionDetails,
+            ApprovalToolInputLabel,
+            ApprovalCommandLabel,
+            ApprovalDetailsTitle,
+            ApprovalDetailsSourceLabel,
+            ApprovalDetailsRunLabel,
+            ApprovalDetailsExecutionLabel,
+            ApprovalDetailsCommandBlockLabel,
+            ApprovalDetailsRedactionLabel,
+            ApprovalDetailsProviderRequestLabel,
+            ApprovalDetailsToolUseLabel,
+            ApprovalDetailsDefaultDenyLine,
+            ApprovalDetailsRequestLabel,
+            ApprovalDetailsInputLabel,
+            ApprovalDetailsBashCommandSubject,
+            ApprovalDetailsShellCommandSubject,
+            ApprovalDetailsToolSubject,
+            ApprovalDetailsPendingValue,
+            ApprovalDetailsNoneValue,
+            ApprovalDetailsNotApplicableValue,
+            ApprovalAssessmentSummaryLine,
+            ApprovalAssessmentReasonLine,
+            ApprovalJournalTitle,
+            ApprovalJournalDecisionCount,
+            ApprovalJournalEmptyBody,
+            ApprovalJournalActorLabel,
+            ApprovalJournalPreviewHashLabel,
+            ApprovalJournalSubjectLabel,
+            ApprovalJournalPreviewLabel,
+            ApprovalRiskSuffix,
+            ApprovalQueueCompactLine,
+            ApprovalQueueFullLine,
+            ApprovalQueueNextSuffix,
+            ApprovalSubjectLabel,
+            ApprovalNextLabel,
+            ApprovalKeysPrefix,
+            ApprovalKeysText,
+            ApprovalExecutableToolPolicy,
+            ApprovalExecutableToolPolicyExtra,
+            ApprovalCommandDefaultPolicy,
+            ApprovalRunShellCommandPrompt,
+            ApprovalRunBashCommandPrompt,
+            ApprovalNotFoundTitle,
+            ApprovalNotFoundBody,
+            ApprovalShellHandoffNotFoundTitle,
+            ApprovalShellHandoffNotFoundBody,
+            ApprovalShellHandoffBlockedTitle,
+            ApprovalShellHandoffBlockedFooter,
+            ApprovalShellHandoffValidationEmptyCommand,
+            ApprovalShellHandoffValidationMultilineCommand,
+            ApprovalShellHandoffValidationControlCharacter,
+            ApprovalShellHandoffValidationEmptyPreview,
+            ApprovalShellHandoffValidationEmptyApprovalId,
+            ApprovalShellHandoffValidationEmptyRunId,
+            ApprovalShellHandoffSendingTitle,
+            ApprovalShellHandoffSendingBody,
+            ApprovalShellHandoffTimeoutTitle,
+            ApprovalShellHandoffTimeoutExceededBody,
+            ApprovalShellHandoffTimeoutInterruptBody,
+            ApprovalReceiptKindToolRequest,
+            ApprovalReceiptKindShellCommandRequest,
+            ApprovalReceiptKindBashTool,
+            ApprovalReceiptDecisionPending,
+            ApprovalReceiptDecisionApproved,
+            ApprovalReceiptDecisionSentToShell,
+            ApprovalReceiptDecisionProviderNativeAllowed,
+            ApprovalReceiptDecisionApprovedDisplayOnly,
+            ApprovalReceiptDecisionDenied,
+            ApprovalReceiptDecisionCancelled,
+            ApprovalReceiptDecisionBlocked,
+            ApprovalReceiptSubjectBashSentToShell,
+            ApprovalReceiptSubjectBashProviderNative,
+            ApprovalReceiptBashSentToShellMessage,
+            ApprovalReceiptProviderNativeAllowedMessage,
+            ApprovalHookHeading,
+        );
+    };
+}
