+++ src/cosh-ng/crates/cosh-shell/src/i18n/message_id/health.rs
@@ -1,0 +1,102 @@
+macro_rules! health_ids {
+    ($next:ident, $remaining:tt, $($ids:ident,)*) => {
+        $next!(
+            $remaining,
+            $($ids,)*
+            HealthBannerTitle,
+            HealthStartupLabel,
+            HealthBannerTryLabel,
+            HealthBannerFindingLabel,
+            HealthBannerEvidenceLabel,
+            HealthBannerMoreFindingsLabel,
+            HealthBannerUnavailableLabel,
+            HealthBannerFindingsSection,
+            HealthBannerSuggestedPromptSection,
+            HealthBannerSuggestedPromptIntro,
+            HealthSeverityOk,
+            HealthSeverityWarning,
+            HealthSeverityCritical,
+            HealthSeverityDegraded,
+            HealthSeverityUnavailable,
+            HealthMetricCpu,
+            HealthMetricCpuLoadPerCore,
+            HealthMetricLoad1mShort,
+            HealthMetricLoad5mShort,
+            HealthMetricLoadValue,
+            HealthMetricCpuPerCoreUnit,
+            HealthMetricCpuUsed,
+            HealthMetricHost,
+            HealthMetricMemory,
+            HealthMetricMemoryAvailable,
+            HealthMetricMemoryUsed,
+            HealthMetricSwap,
+            HealthMetricSwapUsed,
+            HealthMetricDisk,
+            HealthMetricDiskUsed,
+            HealthMetricDiskMountUsed,
+            HealthMetricPressure,
+            HealthMetricLevels,
+            HealthMetricSignal,
+            HealthMetricService,
+            HealthEvidenceDiskAvailable,
+            HealthEvidenceMount,
+            HealthEvidenceOomKilledProcess,
+            HealthEvidenceOomCgroup,
+            HealthEvidenceOomOneHourCount,
+            HealthEvidenceOomTwentyFourHourCount,
+            HealthEvidenceOomVictimKilledAgo,
+            HealthEvidenceOomVictimKilled,
+            HealthEvidenceOomAge,
+            HealthOomScopeMemcg,
+            HealthOomScopeHost,
+            HealthOomScopeCpuset,
+            HealthOomScopeMemoryPolicy,
+            HealthOomScopeUnknown,
+            HealthTryAnalyzeMemoryPressure,
+            HealthTryCheckSwapPressure,
+            HealthTryCheckRecentOom,
+            HealthTryInspectDiskUsage,
+            HealthTryInspectServiceStatus,
+            HealthTryInspectHighLoad,
+            HealthTryInspectProcessMemory,
+            HealthTryReviewUnavailableChecks,
+            HealthPromptAnalyzeMemoryPressure,
+            HealthPromptCheckSwapPressure,
+            HealthPromptCheckRecentOom,
+            HealthPromptInspectDiskUsage,
+            HealthPromptInspectServiceStatus,
+            HealthPromptInspectHighLoad,
+            HealthPromptInspectProcessMemory,
+            HealthPromptReviewUnavailableChecks,
+            HealthInsightMemoryAvailableLow,
+            HealthInsightDiskHigh,
+            HealthInsightRecentOom,
+            HealthInsightCpuLoadHigh,
+            HealthInsightSwapPressure,
+            HealthInsightServiceState,
+            HealthInsightGeneric,
+            HealthUnavailablePermissionDenied,
+            HealthUnavailableCommandMissing,
+            HealthUnavailableTimeout,
+            HealthUnavailableUnsupported,
+            HealthUnavailableParseError,
+            HealthFindingPlatformUnsupported,
+            HealthFindingCoreCollectorUnavailable,
+            HealthFindingCpuLoadHigh,
+            HealthFindingMemoryAvailableLow,
+            HealthFindingSwapPressure,
+            HealthFindingDiskHigh,
+            HealthFindingRecentOom,
+            HealthFindingKernelPanic,
+            HealthFindingServiceFailed,
+            HealthFindingServiceInactive,
+            HealthTryReasonMemoryLow,
+            HealthTryReasonSwapWithContext,
+            HealthTryReasonRecentOom,
+            HealthTryReasonDiskHigh,
+            HealthTryReasonServiceState,
+            HealthTryReasonHighLoad,
+            HealthTryReasonMissingCoreCheck,
+        );
+    };
+}
