+++ src/cosh-ng/crates/cosh-shell/src/i18n/message_id/hook_details.rs
@@ -1,0 +1,61 @@
+macro_rules! hook_details_ids {
+    ($next:ident, $remaining:tt, $($ids:ident,)*) => {
+        $next!(
+            $remaining,
+            $($ids,)*
+            HookFindingTitle,
+            HookFindingFooter,
+            HookFindingMarkdownTitle,
+            HookFindingMarkdownHookLine,
+            HookFindingMarkdownSeverityLine,
+            HookFindingMarkdownFindingLine,
+            HookFindingMarkdownSuggestionLine,
+            HookFindingMarkdownRelatedTitle,
+            HookFindingMarkdownRelatedLine,
+            HookFindingMarkdownAgentFollowUpLine,
+            HookHintTitle,
+            HookHintNotFoundBody,
+            HookHintNotFoundFooter,
+            HookHintNoFindingBody,
+            HookHintBlockUnavailableBody,
+            HookHintIgnoredTitle,
+            HookHintIgnoredBody,
+            HookHintIgnoredFooter,
+            HookHintUsageTitle,
+            HookHintUsageBody,
+            HookFindingDetailsTitle,
+            HookConsultationHookLabel,
+            HookConsultationConfidenceReasonLine,
+            HookConsultationFindingLine,
+            HookConsultationSuggestionLine,
+            HookConsultationAnalyzeAction,
+            HookConsultationIgnoreAction,
+            HookDetailsConfidenceLine,
+            HookDetailsUserInterestLine,
+            HookDetailsReasonLookupIntent,
+            HookDetailsReasonPipelineIntent,
+            HookDetailsReasonScriptIntent,
+            HookDetailsReasonWrapperLowConfidence,
+            HookDetailsReasonInteractiveIntent,
+            HookDetailsReasonActiveRunDeferred,
+            HookDetailsReasonUserContinuedInput,
+            HookDetailsReasonNonDiagnosticSuccessCommand,
+            HookDetailsReasonFeedbackNoisy,
+            HookDetailsReasonIgnoredSameFinding,
+            HookDetailsReasonSameCardAlreadyRendered,
+            HookDetailsReasonInterruptionBudget,
+            HookDetailsReasonLowConfidence,
+            HookDetailsReasonDiagnosticIntent,
+            HookDetailsReasonOtherIntent,
+            HookDetailsTopicLine,
+            HookDetailsOriginLine,
+            HookDetailsSuppressionKeyLine,
+            HookDetailsOutputRefLine,
+            HookDetailsCreatedAtLine,
+            HookDetailsPromptHintLine,
+            HookDetailsRecommendedSkillLine,
+            HookDetailsReadOnlyCliHintLine,
+            HookDetailsFooter,
+        );
+    };
+}
