+++ src/cosh-ng/crates/cosh-shell/src/i18n/message_id/hooks.rs
@@ -1,0 +1,74 @@
+macro_rules! hooks_command_ids {
+    ($next:ident, $remaining:tt, $($ids:ident,)*) => {
+        $next!(
+            $remaining,
+            $($ids,)*
+            SlashHooksRegisteredTitle,
+            SlashHooksNoHooksBody,
+            SlashHooksStatusCountLine,
+            SlashHooksStatusSourcesLine,
+            SlashHooksStatusProjectTrustLine,
+            SlashHooksFooterCount,
+            SlashHooksFooterMutedTargets,
+            SlashHooksTargetMutedTitle,
+            SlashHooksTargetMutedBody,
+            SlashHooksTargetMutedFooter,
+            SlashHooksTargetUnmutedTitle,
+            SlashHooksTargetUnmutedBody,
+            SlashHooksTargetNotMutedBody,
+            SlashHooksEnabledTitle,
+            SlashHooksEnabledBody,
+            SlashHooksDisabledTitle,
+            SlashHooksDisabledBody,
+            SlashHooksHistoryTitle,
+            SlashHooksHistoryEmptyBody,
+            SlashHooksHistoryFooter,
+            SlashHooksEventsTitle,
+            SlashHooksEventsEmptyBody,
+            SlashHooksEventsFooter,
+            SlashHooksUsageTitle,
+            SlashHooksUsageListLine,
+            SlashHooksUsageHistoryLine,
+            SlashHooksUsageEventsLine,
+            SlashHooksUsageAnalyzeLine,
+            SlashHooksUsageIgnoreLine,
+            SlashHooksUsageDetailsLine,
+            SlashHooksUsageFeedbackLine,
+            SlashHooksUsageClearFeedbackLine,
+            SlashHooksUsageMuteLine,
+            SlashHooksUsageUnmuteLine,
+            SlashHooksUsageTrustProjectLine,
+            SlashHooksUsageUntrustProjectLine,
+            SlashHooksUsageClearProjectTrustLine,
+            SlashHooksUsageEnableLine,
+            SlashHooksUsageDisableLine,
+            SlashHooksProjectTrustedTitle,
+            SlashHooksProjectUntrustedTitle,
+            SlashHooksProjectTrustNoHooksBody,
+            SlashHooksProjectTrustedBody,
+            SlashHooksProjectUntrustedBody,
+            SlashHooksProjectTrustNoChangeFooter,
+            SlashHooksProjectTrustPersistedFooter,
+            SlashHooksProjectTrustRemovedFooter,
+            SlashHooksProjectTrustPersistenceFailedFooter,
+            SlashHooksProjectTrustClearedTitle,
+            SlashHooksProjectTrustClearedBody,
+            SlashHooksProjectTrustClearedFooter,
+            SlashHooksProjectTrustClearFailedFooter,
+            SlashHooksFeedbackUsageBody,
+            SlashHooksFeedbackTitle,
+            SlashHooksFeedbackFindingNotFoundBody,
+            SlashHooksFeedbackFindingNotFoundFooter,
+            SlashHooksFeedbackRecordedTitle,
+            SlashHooksFeedbackRecordedBody,
+            SlashHooksFeedbackHookLine,
+            SlashHooksFeedbackPolicyKeyLine,
+            SlashHooksFeedbackPersistedFooter,
+            SlashHooksFeedbackPersistenceFailedFooter,
+            SlashHooksFeedbackClearedTitle,
+            SlashHooksFeedbackClearedBody,
+            SlashHooksFeedbackClearedFooter,
+            SlashHooksFeedbackClearFailedFooter,
+        );
+    };
+}
