+++ src/cosh-ng/crates/cosh-shell/src/i18n/message_id/modes.rs
@@ -1,0 +1,66 @@
+macro_rules! legacy_approval_mode_ids {
+    ($next:ident, $remaining:tt, $($ids:ident,)*) => {
+        $next!(
+            $remaining,
+            $($ids,)*
+            ApprovalModeRemovedBody,
+            ApprovalModeRemovedFooter,
+        );
+    };
+}
+
+macro_rules! mode_ids {
+    ($next:ident, $remaining:tt, $($ids:ident,)*) => {
+        $next!(
+            $remaining,
+            $($ids,)*
+            ModeTitle,
+            ModesTitle,
+            ModeApprovalLine,
+            ModeAnalysisLine,
+            ModeSummaryFooter,
+            ModeRemovedTitle,
+            ModeRemovedBody,
+            ModeRemovedFooter,
+            ModeLanguageBody,
+            ModeLanguageFooter,
+            ModeUnknownBody,
+            ModeUnknownFooter,
+            ApprovalModeTitle,
+            ApprovalModeSetBody,
+            ApprovalModeUnknownBody,
+            ApprovalModeUsageFooter,
+            ApprovalModeRecommendFooter,
+            ApprovalModeAutoFooter,
+            ApprovalModeTrustFooter,
+            ApprovalModeTrustConfirmationTitle,
+            ApprovalModeTrustConfirmationBody,
+            ApprovalModeTrustConfirmationCommandBody,
+            ApprovalModeTrustConfirmationFooter,
+            ApprovalModeCardTitle,
+            ApprovalModeCardCurrentLine,
+            ApprovalModeCardRecommendLine,
+            ApprovalModeCardAutoLine,
+            ApprovalModeCardTrustLine,
+            ApprovalModeCardFooter,
+            ApprovalModeRemainsBody,
+            ApprovalModeCancelBody,
+            ApprovalModeCancelFooter,
+            AnalysisModeTitle,
+            AnalysisModeCurrentBody,
+            AnalysisModeSetBody,
+            AnalysisModeUnknownBody,
+            AnalysisModeUsageFooter,
+            AnalysisModeSmartFooter,
+            AnalysisModeAutoFooter,
+            AnalysisModeManualFooter,
+            AnalysisModeCardSmartLine,
+            AnalysisModeCardAutoLine,
+            AnalysisModeCardManualLine,
+            AnalysisModeCardFooter,
+            AnalysisModeRemainsBody,
+            AnalysisModeCancelBody,
+            AnalysisModeCancelFooter,
+        );
+    };
+}
