+++ src/cosh-ng/crates/cosh-shell/src/i18n/message_id/question.rs
@@ -1,0 +1,22 @@
+macro_rules! question_ids {
+    ($next:ident, $remaining:tt, $($ids:ident,)*) => {
+        $next!(
+            $remaining,
+            $($ids,)*
+            QuestionTitle,
+            QuestionDefaultPrompt,
+            QuestionAnswerLabel,
+            QuestionSelectOneLabel,
+            QuestionSelectMultipleLabel,
+            QuestionOtherEmptyLabel,
+            QuestionKeysPrefix,
+            QuestionInstructionMoveTypeSend,
+            QuestionInstructionMoveToggleSend,
+            QuestionInstructionMoveSend,
+            QuestionInstructionTypeSend,
+            QuestionInstructionNoAnswer,
+            QuestionNoPendingTitle,
+            QuestionNoPendingBody,
+        );
+    };
+}
