+++ src/cosh-ng/crates/cosh-shell/src/i18n/message_id/startup.rs
@@ -1,0 +1,19 @@
+macro_rules! startup_ids {
+    ($next:ident, $remaining:tt, $($ids:ident,)*) => {
+        $next!(
+            $remaining,
+            $($ids,)*
+            StartupTitle,
+            StartupAdapterLine,
+            StartupCwdLine,
+            StartupCommandsLine,
+            StartupHooksNoneSummary,
+            StartupHooksCompletedSummary,
+            StartupHooksFindingsHeading,
+            StartupHooksRustProjectFinding,
+            StartupHooksNoFindings,
+            StartupHooksReadOnlyNote,
+            StartupSwitchHint,
+        );
+    };
+}
