--- src/cosh-ng/crates/cosh-shell/src/i18n/message_id_all.rs
@@ -1,700 +1,0 @@
-use super::MessageId;
-
-impl MessageId {
-    pub const ALL: &'static [MessageId] = &[
-        MessageId::StartupTitle,
-        MessageId::StartupAdapterLine,
-        MessageId::StartupCwdLine,
-        MessageId::StartupCommandsLine,
-        MessageId::StartupHooksNoneSummary,
-        MessageId::StartupHooksCompletedSummary,
-        MessageId::StartupHooksFindingsHeading,
-        MessageId::StartupHooksRustProjectFinding,
-        MessageId::StartupHooksNoFindings,
-        MessageId::StartupHooksReadOnlyNote,
-        MessageId::StartupSwitchHint,
-        MessageId::HelpTitle,
-        MessageId::HelpFooter,
-        MessageId::HelpGroupConfig,
-        MessageId::HelpGroupModes,
-        MessageId::HelpGroupHooks,
-        MessageId::HelpSummaryHelp,
-        MessageId::HelpSummaryAuth,
-        MessageId::HelpSummaryConfig,
-        MessageId::HelpSummaryRecommendations,
-        MessageId::HelpSummaryModeApproval,
-        MessageId::HelpSummaryModeAnalysis,
-        MessageId::HelpSummaryAgent,
-        MessageId::HelpSummaryExplain,
-        MessageId::HelpSummaryCancel,
-        MessageId::HelpSummaryDetails,
-        MessageId::HelpSummaryAudit,
-        MessageId::HelpSummaryHooks,
-        MessageId::HelpSummarySelect,
-        MessageId::HelpSummaryCopy,
-        MessageId::HelpSummaryDebug,
-        MessageId::HelpSummaryClear,
-        MessageId::HelpSummaryShell,
-        MessageId::HelpSummaryApprovalModeRemoved,
-        MessageId::SlashHintTitle,
-        MessageId::SlashHintPrefix,
-        MessageId::SlashHintCurrentMode,
-        MessageId::SlashHintFooter,
-        MessageId::SlashUnknownTitle,
-        MessageId::SlashUnknownBody,
-        MessageId::SlashUnknownSuggestionBody,
-        MessageId::SlashUnknownFooter,
-        MessageId::SlashInfoAuditTitle,
-        MessageId::SlashInfoAuditApprovalsBody,
-        MessageId::SlashInfoAuditActivityBody,
-        MessageId::SlashInfoAuditFooter,
-        MessageId::SlashInfoConfigTitle,
-        MessageId::SlashInfoConfigLanguageLine,
-        MessageId::SlashInfoConfigLanguageEffectiveLine,
-        MessageId::SlashInfoConfigPathLine,
-        MessageId::SlashInfoConfigDebugActivityLine,
-        MessageId::SlashInfoConfigAnalysisStrategyLine,
-        MessageId::SlashInfoConfigRenderFallbackLine,
-        MessageId::SlashInfoConfigFooter,
-        MessageId::ConfigInvalidLanguageBody,
-        MessageId::ConfigSupportedLanguagesFooter,
-        MessageId::ConfigUnknownKeyBody,
-        MessageId::ConfigHomeMissingBody,
-        MessageId::ConfigHomeMissingFooter,
-        MessageId::ConfigUnchangedTitle,
-        MessageId::ConfigNoFileChangedBody,
-        MessageId::ConfigSavedTitle,
-        MessageId::ConfigSavedValueLine,
-        MessageId::ConfigCurrentSessionLanguageLine,
-        MessageId::ConfigSavedFooter,
-        MessageId::ConfigSaveFailedTitle,
-        MessageId::ConfigSaveFailedBody,
-        MessageId::ConfigSavePromptTitle,
-        MessageId::ConfigFileLine,
-        MessageId::ConfigPendingChangeLine,
-        MessageId::ConfigSaveButton,
-        MessageId::ConfigCancelButton,
-        MessageId::ConfigApplyKeysFooter,
-        MessageId::ConfigLanguageTitle,
-        MessageId::ConfigLanguageAutoLine,
-        MessageId::ConfigLanguageEnLine,
-        MessageId::ConfigLanguageZhLine,
-        MessageId::ConfigLanguageKeysFooter,
-        MessageId::SlashHooksRegisteredTitle,
-        MessageId::SlashHooksNoHooksBody,
-        MessageId::SlashHooksStatusCountLine,
-        MessageId::SlashHooksStatusSourcesLine,
-        MessageId::SlashHooksStatusProjectTrustLine,
-        MessageId::SlashHooksFooterCount,
-        MessageId::SlashHooksFooterMutedTargets,
-        MessageId::SlashHooksTargetMutedTitle,
-        MessageId::SlashHooksTargetMutedBody,
-        MessageId::SlashHooksTargetMutedFooter,
-        MessageId::SlashHooksTargetUnmutedTitle,
-        MessageId::SlashHooksTargetUnmutedBody,
-        MessageId::SlashHooksTargetNotMutedBody,
-        MessageId::SlashHooksEnabledTitle,
-        MessageId::SlashHooksEnabledBody,
-        MessageId::SlashHooksDisabledTitle,
-        MessageId::SlashHooksDisabledBody,
-        MessageId::SlashHooksHistoryTitle,
-        MessageId::SlashHooksHistoryEmptyBody,
-        MessageId::SlashHooksHistoryFooter,
-        MessageId::SlashHooksEventsTitle,
-        MessageId::SlashHooksEventsEmptyBody,
-        MessageId::SlashHooksEventsFooter,
-        MessageId::SlashHooksUsageTitle,
-        MessageId::SlashHooksUsageListLine,
-        MessageId::SlashHooksUsageHistoryLine,
-        MessageId::SlashHooksUsageEventsLine,
-        MessageId::SlashHooksUsageAnalyzeLine,
-        MessageId::SlashHooksUsageIgnoreLine,
-        MessageId::SlashHooksUsageDetailsLine,
-        MessageId::SlashHooksUsageFeedbackLine,
-        MessageId::SlashHooksUsageClearFeedbackLine,
-        MessageId::SlashHooksUsageMuteLine,
-        MessageId::SlashHooksUsageUnmuteLine,
-        MessageId::SlashHooksUsageTrustProjectLine,
-        MessageId::SlashHooksUsageUntrustProjectLine,
-        MessageId::SlashHooksUsageClearProjectTrustLine,
-        MessageId::SlashHooksUsageEnableLine,
-        MessageId::SlashHooksUsageDisableLine,
-        MessageId::SlashHooksProjectTrustedTitle,
-        MessageId::SlashHooksProjectUntrustedTitle,
-        MessageId::SlashHooksProjectTrustNoHooksBody,
-        MessageId::SlashHooksProjectTrustedBody,
-        MessageId::SlashHooksProjectUntrustedBody,
-        MessageId::SlashHooksProjectTrustNoChangeFooter,
-        MessageId::SlashHooksProjectTrustPersistedFooter,
-        MessageId::SlashHooksProjectTrustRemovedFooter,
-        MessageId::SlashHooksProjectTrustPersistenceFailedFooter,
-        MessageId::SlashHooksProjectTrustClearedTitle,
-        MessageId::SlashHooksProjectTrustClearedBody,
-        MessageId::SlashHooksProjectTrustClearedFooter,
-        MessageId::SlashHooksProjectTrustClearFailedFooter,
-        MessageId::SlashHooksFeedbackUsageBody,
-        MessageId::SlashHooksFeedbackTitle,
-        MessageId::SlashHooksFeedbackFindingNotFoundBody,
-        MessageId::SlashHooksFeedbackFindingNotFoundFooter,
-        MessageId::SlashHooksFeedbackRecordedTitle,
-        MessageId::SlashHooksFeedbackRecordedBody,
-        MessageId::SlashHooksFeedbackHookLine,
-        MessageId::SlashHooksFeedbackPolicyKeyLine,
-        MessageId::SlashHooksFeedbackPersistedFooter,
-        MessageId::SlashHooksFeedbackPersistenceFailedFooter,
-        MessageId::SlashHooksFeedbackClearedTitle,
-        MessageId::SlashHooksFeedbackClearedBody,
-        MessageId::SlashHooksFeedbackClearedFooter,
-        MessageId::SlashHooksFeedbackClearFailedFooter,
-        MessageId::DebugSessionTitle,
-        MessageId::DebugAdapterLine,
-        MessageId::DebugProviderInvocationLine,
-        MessageId::DebugProviderCommittedSessionLine,
-        MessageId::DebugActiveRunLine,
-        MessageId::DebugQueuedRunsLine,
-        MessageId::DebugProviderPendingSessionLine,
-        MessageId::DebugProviderInitializeSeenLine,
-        MessageId::DebugHostExecutedShellResultLine,
-        MessageId::DebugSelectedShellExecutionPathLine,
-        MessageId::DebugLatestProviderRequestLine,
-        MessageId::DebugLatestToolUseLine,
-        MessageId::DebugLatestRecoveryStatusLine,
-        MessageId::DebugLatestRecoveryReasonLine,
-        MessageId::DebugEvidenceAccessLine,
-        MessageId::DebugEvidenceToolRegisteredLine,
-        MessageId::DebugEvidenceNamespaceLine,
-        MessageId::DebugEvidenceLedgerCountLine,
-        MessageId::DebugLatestShellOutputReadLine,
-        MessageId::DebugUnknownTargetBody,
-        MessageId::DebugUnknownTargetFooter,
-        MessageId::CommandRemovedTitle,
-        MessageId::ApprovalModeRemovedBody,
-        MessageId::ApprovalModeRemovedFooter,
-        MessageId::RemovedDecisionCommandBody,
-        MessageId::RemovedApprovalDecisionFooter,
-        MessageId::RemovedQuestionAnswerFooter,
-        MessageId::ModeTitle,
-        MessageId::ModesTitle,
-        MessageId::ModeApprovalLine,
-        MessageId::ModeAnalysisLine,
-        MessageId::ModeSummaryFooter,
-        MessageId::ModeRemovedTitle,
-        MessageId::ModeRemovedBody,
-        MessageId::ModeRemovedFooter,
-        MessageId::ModeLanguageBody,
-        MessageId::ModeLanguageFooter,
-        MessageId::ModeUnknownBody,
-        MessageId::ModeUnknownFooter,
-        MessageId::ApprovalModeTitle,
-        MessageId::ApprovalModeSetBody,
-        MessageId::ApprovalModeUnknownBody,
-        MessageId::ApprovalModeUsageFooter,
-        MessageId::ApprovalModeRecommendFooter,
-        MessageId::ApprovalModeAutoFooter,
-        MessageId::ApprovalModeTrustFooter,
-        MessageId::ApprovalModeTrustConfirmationTitle,
-        MessageId::ApprovalModeTrustConfirmationBody,
-        MessageId::ApprovalModeTrustConfirmationCommandBody,
-        MessageId::ApprovalModeTrustConfirmationFooter,
-        MessageId::ApprovalModeCardTitle,
-        MessageId::ApprovalModeCardCurrentLine,
-        MessageId::ApprovalModeCardRecommendLine,
-        MessageId::ApprovalModeCardAutoLine,
-        MessageId::ApprovalModeCardTrustLine,
-        MessageId::ApprovalModeCardFooter,
-        MessageId::ApprovalModeRemainsBody,
-        MessageId::ApprovalModeCancelBody,
-        MessageId::ApprovalModeCancelFooter,
-        MessageId::AnalysisModeTitle,
-        MessageId::AnalysisModeCurrentBody,
-        MessageId::AnalysisModeSetBody,
-        MessageId::AnalysisModeUnknownBody,
-        MessageId::AnalysisModeUsageFooter,
-        MessageId::AnalysisModeSmartFooter,
-        MessageId::AnalysisModeAutoFooter,
-        MessageId::AnalysisModeManualFooter,
-        MessageId::AnalysisModeCardSmartLine,
-        MessageId::AnalysisModeCardAutoLine,
-        MessageId::AnalysisModeCardManualLine,
-        MessageId::AnalysisModeCardFooter,
-        MessageId::AnalysisModeRemainsBody,
-        MessageId::AnalysisModeCancelBody,
-        MessageId::AnalysisModeCancelFooter,
-        MessageId::AgentThinking,
-        MessageId::AgentThinkingElapsed,
-        MessageId::AgentRecoveryTitle,
-        MessageId::AgentRecoveryFreshTurnBody,
-        MessageId::AgentRecoveryContinuityBody,
-        MessageId::AgentStatusTitle,
-        MessageId::AgentStillWorking,
-        MessageId::AgentStatusFooter,
-        MessageId::AgentStatusStarting,
-        MessageId::AgentStatusWaitingBackend,
-        MessageId::AgentStatusThinking,
-        MessageId::AgentStatusPreparingModelSession,
-        MessageId::AgentStatusStartingModelBackend,
-        MessageId::AgentStatusModelInitialized,
-        MessageId::AgentStatusModelStatus,
-        MessageId::AgentStatusAnalysisCompleted,
-        MessageId::AgentStatusAnalysisReturnedError,
-        MessageId::AgentStatusStreaming,
-        MessageId::AgentStatusReceivingResponse,
-        MessageId::AgentStatusApproval,
-        MessageId::AgentStatusWaitingApprovalTool,
-        MessageId::AgentStatusQuestion,
-        MessageId::AgentStatusWaitingUserAnswer,
-        MessageId::AgentStatusWaitingApprovalCommand,
-        MessageId::AgentStatusTool,
-        MessageId::AgentStatusCapturingToolOutput,
-        MessageId::AgentStatusToolCompleted,
-        MessageId::AgentStatusCompleted,
-        MessageId::AgentStatusFailed,
-        MessageId::AgentStatusCancelled,
-        MessageId::AgentStatusRunningApprovedProviderTool,
-        MessageId::AgentProviderTimeoutDroppedQueuedBody,
-        MessageId::AgentCancellationRequestedTitle,
-        MessageId::AgentCancellationRequestedBody,
-        MessageId::AgentCancelledReasonLabel,
-        MessageId::AgentCancelledUserRequestedReason,
-        MessageId::AgentResponseTitle,
-        MessageId::AgentGovernanceTitle,
-        MessageId::AgentGovernanceStatusLine,
-        MessageId::AgentGovernanceReasonLine,
-        MessageId::AgentGovernanceSummaryLine,
-        MessageId::AgentGovernanceErrorLine,
-        MessageId::AgentGovernanceToolOutputLine,
-        MessageId::AgentGovernanceToolCompletedLine,
-        MessageId::AgentGovernanceApprovalRequiredLine,
-        MessageId::AgentGovernanceShellCommandSubject,
-        MessageId::AgentGovernanceBashCommandSubject,
-        MessageId::AgentGovernanceToolSubject,
-        MessageId::AgentGovernanceBlockedUserApprovalLine,
-        MessageId::AgentGovernanceQuestionLine,
-        MessageId::AgentRecommendedCommandsLabel,
-        MessageId::InterceptNoticeTitle,
-        MessageId::InterceptNoticeBody,
-        MessageId::InterceptNoticeFooter,
-        MessageId::FailedAnalysisCancelledTitle,
-        MessageId::FailedAnalysisCancelledBody,
-        MessageId::FailedAnalysisCancelNoActiveBody,
-        MessageId::FailedAnalysisCancelledFooter,
-        MessageId::AnalysisSkippedTitle,
-        MessageId::AnalysisSkippedBody,
-        MessageId::AnalysisSkippedFooter,
-        MessageId::HookAutoAnalyzedTitle,
-        MessageId::HookAutoAnalyzedBody,
-        MessageId::HookAutoAnalyzedFooter,
-        MessageId::InsightLabel,
-        MessageId::InsightCommandTypoSummary,
-        MessageId::InsightPermissionDeniedSummary,
-        MessageId::InsightBuildOrTestFailureSummary,
-        MessageId::InsightRuntimeExceptionSummary,
-        MessageId::InsightAbnormalSignalSummary,
-        MessageId::InsightMemoryPressureSummary,
-        MessageId::InsightHighMemoryProcessSummary,
-        MessageId::InsightHighMemoryProcessGenericSummary,
-        MessageId::InsightMemoryRootCauseSummary,
-        MessageId::InsightMemoryRootCauseGenericSummary,
-        MessageId::InsightPermissionDeniedPrompt,
-        MessageId::InsightBuildOrTestFailurePrompt,
-        MessageId::InsightRuntimeExceptionPrompt,
-        MessageId::InsightAbnormalSignalPrompt,
-        MessageId::InsightMemoryPressurePrompt,
-        MessageId::InsightHighMemoryProcessPrompt,
-        MessageId::InsightHighMemoryProcessGenericPrompt,
-        MessageId::InsightMemoryRootCausePrompt,
-        MessageId::InsightMemoryRootCauseGenericPrompt,
-        MessageId::InsightShellRewriteFirstUseHint,
-        MessageId::InsightAgentPromptFirstUseHint,
-        MessageId::AgentQueuedTitle,
-        MessageId::AgentQueuedBodyCommand,
-        MessageId::AgentQueuedBodyActive,
-        MessageId::AgentQueuedFooter,
-        MessageId::HookFindingTitle,
-        MessageId::HookFindingFooter,
-        MessageId::HookFindingMarkdownTitle,
-        MessageId::HookFindingMarkdownHookLine,
-        MessageId::HookFindingMarkdownSeverityLine,
-        MessageId::HookFindingMarkdownFindingLine,
-        MessageId::HookFindingMarkdownSuggestionLine,
-        MessageId::HookFindingMarkdownRelatedTitle,
-        MessageId::HookFindingMarkdownRelatedLine,
-        MessageId::HookFindingMarkdownAgentFollowUpLine,
-        MessageId::HookHintTitle,
-        MessageId::HookHintNotFoundBody,
-        MessageId::HookHintNotFoundFooter,
-        MessageId::HookHintNoFindingBody,
-        MessageId::HookHintBlockUnavailableBody,
-        MessageId::HookHintIgnoredTitle,
-        MessageId::HookHintIgnoredBody,
-        MessageId::HookHintIgnoredFooter,
-        MessageId::HookHintUsageTitle,
-        MessageId::HookHintUsageBody,
-        MessageId::HookFindingDetailsTitle,
-        MessageId::HookConsultationHookLabel,
-        MessageId::HookConsultationConfidenceReasonLine,
-        MessageId::HookConsultationAnalyzeAction,
-        MessageId::HookConsultationIgnoreAction,
-        MessageId::HookDetailsConfidenceLine,
-        MessageId::HookDetailsUserInterestLine,
-        MessageId::HookDetailsReasonLookupIntent,
-        MessageId::HookDetailsReasonPipelineIntent,
-        MessageId::HookDetailsReasonScriptIntent,
-        MessageId::HookDetailsReasonWrapperLowConfidence,
-        MessageId::HookDetailsReasonInteractiveIntent,
-        MessageId::HookDetailsReasonActiveRunDeferred,
-        MessageId::HookDetailsReasonUserContinuedInput,
-        MessageId::HookDetailsReasonNonDiagnosticSuccessCommand,
-        MessageId::HookDetailsReasonFeedbackNoisy,
-        MessageId::HookDetailsReasonIgnoredSameFinding,
-        MessageId::HookDetailsReasonSameCardAlreadyRendered,
-        MessageId::HookDetailsReasonInterruptionBudget,
-        MessageId::HookDetailsReasonLowConfidence,
-        MessageId::HookDetailsReasonDiagnosticIntent,
-        MessageId::HookDetailsReasonOtherIntent,
-        MessageId::HookDetailsTopicLine,
-        MessageId::HookDetailsSuppressionKeyLine,
-        MessageId::HookDetailsOutputRefLine,
-        MessageId::HookDetailsCreatedAtLine,
-        MessageId::HookDetailsPromptHintLine,
-        MessageId::HookDetailsRecommendedSkillLine,
-        MessageId::HookDetailsReadOnlyCliHintLine,
-        MessageId::HookDetailsFooter,
-        MessageId::RuntimeDetailsUnavailableTitle,
-        MessageId::RuntimeDetailsUnavailableBody,
-        MessageId::ActivityTitle,
-        MessageId::ActivityDetailsTitle,
-        MessageId::ActivityRunLabel,
-        MessageId::ActivityDetailLabel,
-        MessageId::ActivitySkillLabel,
-        MessageId::ActivitySkillUpdatedStatus,
-        MessageId::ActivityToolLabel,
-        MessageId::ActivityToolOutputLabel,
-        MessageId::ActivityShellLabel,
-        MessageId::ActivityStatusLoading,
-        MessageId::ActivityStatusLoaded,
-        MessageId::ActivityStatusFailed,
-        MessageId::ActivityStatusCalled,
-        MessageId::ActivityStatusRequested,
-        MessageId::ActivityStatusCaptured,
-        MessageId::ActivityStatusCompleted,
-        MessageId::ActivityStatusError,
-        MessageId::ActivityStatusInterrupted,
-        MessageId::ActivityToolCalledSummary,
-        MessageId::ActivityToolRequestedSummary,
-        MessageId::ActivityToolOutputCapturedSummary,
-        MessageId::ActivityToolNeedsForegroundShellSummary,
-        MessageId::ActivityShellHandoffSentSummary,
-        MessageId::ToolCardReadFileLabel,
-        MessageId::ToolCardWriteFileLabel,
-        MessageId::ToolCardEditFileLabel,
-        MessageId::ToolCardSearchFilesLabel,
-        MessageId::ToolCardFindFilesLabel,
-        MessageId::ToolCardListDirectoryLabel,
-        MessageId::ToolCardShellLabel,
-        MessageId::ToolCardWebFetchLabel,
-        MessageId::ToolCardWebSearchLabel,
-        MessageId::ToolCardSkillLabel,
-        MessageId::ToolCardAgentLabel,
-        MessageId::ToolCardMemoryLabel,
-        MessageId::ToolCardEvidenceLabel,
-        MessageId::ToolCardCustomToolLabel,
-        MessageId::ToolCardCalledStatus,
-        MessageId::ToolCardRequestedStatus,
-        MessageId::ToolCardAutoApprovedStatus,
-        MessageId::ToolCardCapturedStatus,
-        MessageId::ToolCardCompletedStatus,
-        MessageId::ToolCardFailedStatus,
-        MessageId::ToolCardDuplicateStatus,
-        MessageId::ToolCardInterruptedStatus,
-        MessageId::ToolCardReadOnlyIntent,
-        MessageId::ToolCardWriteIntent,
-        MessageId::ToolCardExecuteIntent,
-        MessageId::ToolCardNetworkIntent,
-        MessageId::ToolCardContextIntent,
-        MessageId::ToolCardCustomIntent,
-        MessageId::ToolCardApprovalRequiredAction,
-        MessageId::ToolCardWriteCompletedResult,
-        MessageId::ToolCardEditCompletedResult,
-        MessageId::ToolCardSkillAvailableResult,
-        MessageId::ToolCardShellEvidenceDeliveredResult,
-        MessageId::ToolCardShellEvidenceListResult,
-        MessageId::ToolCardShellEvidenceReadResult,
-        MessageId::ToolCardShellEvidenceAlreadyDeliveredResult,
-        MessageId::ToolCardShellEvidenceFailedResult,
-        MessageId::ToolCardShellEvidenceDuplicateResult,
-        MessageId::ToolCardShellEvidenceMetadataMetric,
-        MessageId::ToolCardOutputCapturedResult,
-        MessageId::ToolCardLinesReturnedResult,
-        MessageId::ToolCardStdoutMetric,
-        MessageId::ToolCardStderrMetric,
-        MessageId::ToolCardTruncatedMetric,
-        MessageId::MarkdownCodeLabel,
-        MessageId::MarkdownCodeWithLanguageLabel,
-        MessageId::MarkdownTableLabel,
-        MessageId::RecommendationTitle,
-        MessageId::RecommendationNextStepTitle,
-        MessageId::AnalysisResultTitle,
-        MessageId::RecommendationEmptyBody,
-        MessageId::RecommendationFooter,
-        MessageId::RecommendationNoSelectableTitle,
-        MessageId::RecommendationNoSelectableBody,
-        MessageId::RecommendationUnavailableTitle,
-        MessageId::RecommendationUnavailableBody,
-        MessageId::RecommendationSelectedTitle,
-        MessageId::RecommendationSelectedBody,
-        MessageId::RecommendationCopiedTitle,
-        MessageId::RecommendationCopiedBody,
-        MessageId::RecommendationInsertTitle,
-        MessageId::RecommendationInsertBody,
-        MessageId::RecommendationDetailsTitle,
-        MessageId::RecommendationDetailsBody,
-        MessageId::RecommendationDisplayOnlyBody,
-        MessageId::RecommendationCopyOnlyBody,
-        MessageId::RecommendationInsertOnlyBody,
-        MessageId::RecommendationDetailsOnlyBody,
-        MessageId::HealthBannerTitle,
-        MessageId::HealthStartupLabel,
-        MessageId::HealthBannerTryLabel,
-        MessageId::HealthBannerFindingLabel,
-        MessageId::HealthBannerEvidenceLabel,
-        MessageId::HealthBannerMoreFindingsLabel,
-        MessageId::HealthBannerUnavailableLabel,
-        MessageId::HealthBannerFindingsSection,
-        MessageId::HealthBannerSuggestedPromptSection,
-        MessageId::HealthBannerSuggestedPromptIntro,
-        MessageId::HealthSeverityOk,
-        MessageId::HealthSeverityWarning,
-        MessageId::HealthSeverityCritical,
-        MessageId::HealthSeverityDegraded,
-        MessageId::HealthSeverityUnavailable,
-        MessageId::HealthMetricCpu,
-        MessageId::HealthMetricCpuLoadPerCore,
-        MessageId::HealthMetricLoad1mShort,
-        MessageId::HealthMetricLoad5mShort,
-        MessageId::HealthMetricLoadValue,
-        MessageId::HealthMetricCpuPerCoreUnit,
-        MessageId::HealthMetricCpuUsed,
-        MessageId::HealthMetricHost,
-        MessageId::HealthMetricMemory,
-        MessageId::HealthMetricMemoryAvailable,
-        MessageId::HealthMetricMemoryUsed,
-        MessageId::HealthMetricSwap,
-        MessageId::HealthMetricSwapUsed,
-        MessageId::HealthMetricDisk,
-        MessageId::HealthMetricDiskUsed,
-        MessageId::HealthMetricDiskMountUsed,
-        MessageId::HealthMetricPressure,
-        MessageId::HealthMetricLevels,
-        MessageId::HealthMetricSignal,
-        MessageId::HealthMetricService,
-        MessageId::HealthEvidenceDiskAvailable,
-        MessageId::HealthEvidenceMount,
-        MessageId::HealthEvidenceOomKilledProcess,
-        MessageId::HealthEvidenceOomCgroup,
-        MessageId::HealthEvidenceOomOneHourCount,
-        MessageId::HealthEvidenceOomTwentyFourHourCount,
-        MessageId::HealthEvidenceOomVictimKilledAgo,
-        MessageId::HealthEvidenceOomVictimKilled,
-        MessageId::HealthEvidenceOomAge,
-        MessageId::HealthOomScopeMemcg,
-        MessageId::HealthOomScopeHost,
-        MessageId::HealthOomScopeCpuset,
-        MessageId::HealthOomScopeMemoryPolicy,
-        MessageId::HealthOomScopeUnknown,
-        MessageId::HealthTryAnalyzeMemoryPressure,
-        MessageId::HealthTryCheckSwapPressure,
-        MessageId::HealthTryCheckRecentOom,
-        MessageId::HealthTryInspectDiskUsage,
-        MessageId::HealthTryInspectServiceStatus,
-        MessageId::HealthTryInspectHighLoad,
-        MessageId::HealthTryInspectProcessMemory,
-        MessageId::HealthTryReviewUnavailableChecks,
-        MessageId::HealthPromptAnalyzeMemoryPressure,
-        MessageId::HealthPromptCheckSwapPressure,
-        MessageId::HealthPromptCheckRecentOom,
-        MessageId::HealthPromptInspectDiskUsage,
-        MessageId::HealthPromptInspectServiceStatus,
-        MessageId::HealthPromptInspectHighLoad,
-        MessageId::HealthPromptInspectProcessMemory,
-        MessageId::HealthPromptReviewUnavailableChecks,
-        MessageId::HealthInsightMemoryAvailableLow,
-        MessageId::HealthInsightDiskHigh,
-        MessageId::HealthInsightRecentOom,
-        MessageId::HealthInsightCpuLoadHigh,
-        MessageId::HealthInsightSwapPressure,
-        MessageId::HealthInsightServiceState,
-        MessageId::HealthInsightGeneric,
-        MessageId::HealthUnavailablePermissionDenied,
-        MessageId::HealthUnavailableCommandMissing,
-        MessageId::HealthUnavailableTimeout,
-        MessageId::HealthUnavailableUnsupported,
-        MessageId::HealthUnavailableParseError,
-        MessageId::HealthFindingPlatformUnsupported,
-        MessageId::HealthFindingCoreCollectorUnavailable,
-        MessageId::HealthFindingCpuLoadHigh,
-        MessageId::HealthFindingMemoryAvailableLow,
-        MessageId::HealthFindingSwapPressure,
-        MessageId::HealthFindingDiskHigh,
-        MessageId::HealthFindingRecentOom,
-        MessageId::HealthFindingKernelPanic,
-        MessageId::HealthFindingServiceFailed,
-        MessageId::HealthFindingServiceInactive,
-        MessageId::HealthTryReasonMemoryLow,
-        MessageId::HealthTryReasonSwapWithContext,
-        MessageId::HealthTryReasonRecentOom,
-        MessageId::HealthTryReasonDiskHigh,
-        MessageId::HealthTryReasonServiceState,
-        MessageId::HealthTryReasonHighLoad,
-        MessageId::HealthTryReasonMissingCoreCheck,
-        MessageId::ToolOutputStdoutCapturedSummary,
-        MessageId::ToolOutputStderrCapturedSummary,
-        MessageId::ToolSummaryExit,
-        MessageId::ToolSummaryBlocked,
-        MessageId::ToolSummaryTimedOut,
-        MessageId::ToolSummaryFailed,
-        MessageId::QuestionTitle,
-        MessageId::QuestionDefaultPrompt,
-        MessageId::QuestionAnswerLabel,
-        MessageId::QuestionSelectOneLabel,
-        MessageId::QuestionSelectMultipleLabel,
-        MessageId::QuestionOtherEmptyLabel,
-        MessageId::QuestionKeysPrefix,
-        MessageId::QuestionInstructionMoveTypeSend,
-        MessageId::QuestionInstructionMoveToggleSend,
-        MessageId::QuestionInstructionMoveSend,
-        MessageId::QuestionInstructionTypeSend,
-        MessageId::QuestionInstructionNoAnswer,
-        MessageId::QuestionNoPendingTitle,
-        MessageId::QuestionNoPendingBody,
-        MessageId::ApprovalTitle,
-        MessageId::ApprovalRequiredTitle,
-        MessageId::ApprovalResolutionApprovedTitle,
-        MessageId::ApprovalResolutionAutoApprovedTitle,
-        MessageId::ApprovalResolutionTrustedTitle,
-        MessageId::ApprovalResolutionDeniedTitle,
-        MessageId::ApprovalResolutionCancelledTitle,
-        MessageId::ApprovalResolutionBlockedTitle,
-        MessageId::ApprovalResolutionDeferredTitle,
-        MessageId::ApprovalActionAllowOnce,
-        MessageId::ApprovalActionAlwaysTrust,
-        MessageId::ApprovalActionDeny,
-        MessageId::ApprovalActionDetails,
-        MessageId::ApprovalToolInputLabel,
-        MessageId::ApprovalCommandLabel,
-        MessageId::ApprovalDetailsTitle,
-        MessageId::ApprovalDetailsSourceLabel,
-        MessageId::ApprovalDetailsRunLabel,
-        MessageId::ApprovalDetailsExecutionLabel,
-        MessageId::ApprovalDetailsCommandBlockLabel,
-        MessageId::ApprovalDetailsRedactionLabel,
-        MessageId::ApprovalDetailsProviderRequestLabel,
-        MessageId::ApprovalDetailsToolUseLabel,
-        MessageId::ApprovalDetailsDefaultDenyLine,
-        MessageId::ApprovalDetailsRequestLabel,
-        MessageId::ApprovalDetailsInputLabel,
-        MessageId::ApprovalDetailsBashCommandSubject,
-        MessageId::ApprovalDetailsShellCommandSubject,
-        MessageId::ApprovalDetailsToolSubject,
-        MessageId::ApprovalDetailsPendingValue,
-        MessageId::ApprovalDetailsNoneValue,
-        MessageId::ApprovalDetailsNotApplicableValue,
-        MessageId::ApprovalAssessmentSummaryLine,
-        MessageId::ApprovalAssessmentReasonLine,
-        MessageId::ApprovalJournalTitle,
-        MessageId::ApprovalJournalDecisionCount,
-        MessageId::ApprovalJournalEmptyBody,
-        MessageId::ApprovalJournalActorLabel,
-        MessageId::ApprovalJournalPreviewHashLabel,
-        MessageId::ApprovalJournalSubjectLabel,
-        MessageId::ApprovalJournalPreviewLabel,
-        MessageId::ApprovalRiskSuffix,
-        MessageId::ApprovalQueueCompactLine,
-        MessageId::ApprovalQueueFullLine,
-        MessageId::ApprovalQueueNextSuffix,
-        MessageId::ApprovalSubjectLabel,
-        MessageId::ApprovalNextLabel,
-        MessageId::ApprovalKeysPrefix,
-        MessageId::ApprovalKeysText,
-        MessageId::ApprovalExecutableToolPolicy,
-        MessageId::ApprovalExecutableToolPolicyExtra,
-        MessageId::ApprovalCommandDefaultPolicy,
-        MessageId::ApprovalRunShellCommandPrompt,
-        MessageId::ApprovalRunBashCommandPrompt,
-        MessageId::ApprovalNotFoundTitle,
-        MessageId::ApprovalNotFoundBody,
-        MessageId::ApprovalShellHandoffNotFoundTitle,
-        MessageId::ApprovalShellHandoffNotFoundBody,
-        MessageId::ApprovalShellHandoffBlockedTitle,
-        MessageId::ApprovalShellHandoffBlockedFooter,
-        MessageId::ApprovalShellHandoffValidationEmptyCommand,
-        MessageId::ApprovalShellHandoffValidationMultilineCommand,
-        MessageId::ApprovalShellHandoffValidationControlCharacter,
-        MessageId::ApprovalShellHandoffValidationEmptyPreview,
-        MessageId::ApprovalShellHandoffValidationEmptyApprovalId,
-        MessageId::ApprovalShellHandoffValidationEmptyRunId,
-        MessageId::ApprovalShellHandoffSendingTitle,
-        MessageId::ApprovalShellHandoffSendingBody,
-        MessageId::ApprovalShellHandoffTimeoutTitle,
-        MessageId::ApprovalShellHandoffTimeoutExceededBody,
-        MessageId::ApprovalShellHandoffTimeoutInterruptBody,
-        MessageId::ApprovalReceiptKindToolRequest,
-        MessageId::ApprovalReceiptKindShellCommandRequest,
-        MessageId::ApprovalReceiptKindBashTool,
-        MessageId::ApprovalReceiptDecisionPending,
-        MessageId::ApprovalReceiptDecisionApproved,
-        MessageId::ApprovalReceiptDecisionSentToShell,
-        MessageId::ApprovalReceiptDecisionProviderNativeAllowed,
-        MessageId::ApprovalReceiptDecisionApprovedDisplayOnly,
-        MessageId::ApprovalReceiptDecisionDenied,
-        MessageId::ApprovalReceiptDecisionCancelled,
-        MessageId::ApprovalReceiptDecisionBlocked,
-        MessageId::ApprovalReceiptSubjectBashSentToShell,
-        MessageId::ApprovalReceiptSubjectBashProviderNative,
-        MessageId::ApprovalReceiptBashSentToShellMessage,
-        MessageId::ApprovalReceiptProviderNativeAllowedMessage,
-        // Session recovery
-        MessageId::HelpGroupSessions,
-        MessageId::HelpSummarySession,
-        MessageId::SessionTitle,
-        MessageId::SessionUnavailableBody,
-        MessageId::SessionBusyBody,
-        MessageId::SessionEmptyBody,
-        MessageId::SessionListFooter,
-        MessageId::SessionStatusTitle,
-        MessageId::SessionShellIdLine,
-        MessageId::SessionProviderIdLine,
-        MessageId::SessionWorkspaceLine,
-        MessageId::SessionRecoveryLine,
-        MessageId::SessionErrorLine,
-        MessageId::SessionEvidenceNotRestoredBody,
-        MessageId::SessionPickerFooter,
-        MessageId::SessionClearConfirmTitle,
-        MessageId::SessionClearConfirmCountLine,
-        MessageId::SessionClearConfirmFooter,
-        MessageId::SessionSelectedTitle,
-        MessageId::SessionSelectedBody,
-        MessageId::SessionErrorTitle,
-        MessageId::SessionClearedTitle,
-        MessageId::SessionClearedBody,
-        MessageId::SessionSkippedBody,
-        MessageId::SessionClearInterruptedBody,
-        MessageId::SessionCancelledTitle,
-        MessageId::SessionCancelledBody,
-        MessageId::SessionUsageBody,
-        MessageId::SessionNotReadyBody,
-        MessageId::SessionProtectedBody,
-        // Registry slash commands
-        MessageId::HelpGroupRegistry,
-        MessageId::HelpSummaryExtensions,
-        MessageId::HelpSummarySkills,
-        MessageId::SlashExtensionsTitle,
-        MessageId::SlashSkillsTitle,
-        MessageId::SlashRegistryUnavailable,
-        MessageId::SlashHooksShellSection,
-        MessageId::SlashHooksAgentSection,
-        MessageId::SlashHooksAgentUnavailable,
-        MessageId::SlashExtensionsEmptyBody,
-        MessageId::SlashSkillsEmptyBody,
-    ];
-}
