+++ src/cosh-ng/crates/cosh-shell/src/i18n/zh/activity.rs
@@ -1,0 +1,92 @@
+use super::MessageId;
+
+pub(super) fn message(id: MessageId) -> Option<&'static str> {
+    Some(match id {
+        MessageId::RuntimeDetailsUnavailableTitle => "详情不可用",
+        MessageId::RuntimeDetailsUnavailableBody => {
+            "{id} 不可用；请对审批或活动 id 使用 Details 操作"
+        }
+        MessageId::ActivityTitle => "活动",
+        MessageId::ActivityDetailsTitle => "活动详情",
+        MessageId::ActivityRunLabel => "运行",
+        MessageId::ActivityDetailLabel => "详情",
+        MessageId::ActivitySkillLabel => "技能",
+        MessageId::ActivitySkillUpdatedStatus => "已更新",
+        MessageId::ActivityToolLabel => "Tool",
+        MessageId::ActivityToolOutputLabel => "Tool 输出",
+        MessageId::ActivityShellLabel => "Shell",
+        MessageId::ActivityStatusLoading => "加载中",
+        MessageId::ActivityStatusLoaded => "已加载",
+        MessageId::ActivityStatusFailed => "失败",
+        MessageId::ActivityStatusCalled => "已调用",
+        MessageId::ActivityStatusRequested => "请求审批",
+        MessageId::ActivityStatusCaptured => "已捕获",
+        MessageId::ActivityStatusCompleted => "已完成",
+        MessageId::ActivityStatusError => "错误",
+        MessageId::ActivityStatusInterrupted => "已中断",
+        MessageId::ActivityToolCalledSummary => "{tool} 已调用：{preview}",
+        MessageId::ActivityToolRequestedSummary => "{tool} 请求审批：{preview}",
+        MessageId::ActivityToolOutputCapturedSummary => "{stream} 已捕获",
+        MessageId::ActivityProviderNativeShellBypassSummary => {
+            "{tool} 已由 provider 自动批准：{preview}"
+        }
+        MessageId::ActivityToolNeedsForegroundShellSummary => {
+            "可能需要前台 shell；[Send to shell] {handoff}"
+        }
+        MessageId::ActivityShellHandoffSentSummary => "{approval} 已发送到 shell",
+        MessageId::ToolCardReadFileLabel => "读取",
+        MessageId::ToolCardWriteFileLabel => "写入",
+        MessageId::ToolCardEditFileLabel => "编辑",
+        MessageId::ToolCardSearchFilesLabel => "搜索",
+        MessageId::ToolCardFindFilesLabel => "查找文件",
+        MessageId::ToolCardListDirectoryLabel => "列目录",
+        MessageId::ToolCardShellLabel => "Shell",
+        MessageId::ToolCardWebFetchLabel => "网页读取",
+        MessageId::ToolCardWebSearchLabel => "网页搜索",
+        MessageId::ToolCardSkillLabel => "技能",
+        MessageId::ToolCardAgentLabel => "Agent",
+        MessageId::ToolCardMemoryLabel => "记忆",
+        MessageId::ToolCardEvidenceLabel => "Shell 证据",
+        MessageId::ToolCardCustomToolLabel => "自定义工具",
+        MessageId::ToolCardCalledStatus => "已调用",
+        MessageId::ToolCardRequestedStatus => "请求审批",
+        MessageId::ToolCardAutoApprovedStatus => "自动批准",
+        MessageId::ToolCardCapturedStatus => "已捕获",
+        MessageId::ToolCardCompletedStatus => "已完成",
+        MessageId::ToolCardFailedStatus => "失败",
+        MessageId::ToolCardDuplicateStatus => "重复请求",
+        MessageId::ToolCardInterruptedStatus => "已中断",
+        MessageId::ToolCardReadOnlyIntent => "只读操作",
+        MessageId::ToolCardWriteIntent => "将修改工作区状态",
+        MessageId::ToolCardExecuteIntent => "将运行命令",
+        MessageId::ToolCardNetworkIntent => "将访问网络",
+        MessageId::ToolCardContextIntent => "将更新上下文",
+        MessageId::ToolCardCustomIntent => "自定义工具调用",
+        MessageId::ToolCardApprovalRequiredAction => "需要审批",
+        MessageId::ToolCardWriteCompletedResult => "写入完成",
+        MessageId::ToolCardEditCompletedResult => "编辑完成",
+        MessageId::ToolCardSkillAvailableResult => "指令已可用",
+        MessageId::ToolCardShellEvidenceDeliveredResult => "正文未重复输出",
+        MessageId::ToolCardShellEvidenceListResult => "命令历史已交给 Agent",
+        MessageId::ToolCardShellEvidenceReadResult => "Shell 输出片段已交给 Agent",
+        MessageId::ToolCardShellEvidenceAlreadyDeliveredResult => "最近输出已交付，正文不重复显示",
+        MessageId::ToolCardShellEvidenceFailedResult => "Shell 证据不可用",
+        MessageId::ToolCardShellEvidenceDuplicateResult => "Provider 重复发送了相同 Shell 证据请求",
+        MessageId::ToolCardShellEvidenceMetadataMetric => "证据元数据: {count} 行",
+        MessageId::ToolCardOutputCapturedResult => "输出已捕获",
+        MessageId::ToolCardLinesReturnedResult => "返回 {count} 行",
+        MessageId::ToolCardStdoutMetric => "stdout: {count} 行",
+        MessageId::ToolCardStderrMetric => "stderr: {count} 行",
+        MessageId::ToolCardTruncatedMetric => "已截断",
+        MessageId::MarkdownCodeLabel => "代码",
+        MessageId::MarkdownCodeWithLanguageLabel => "代码: {language}",
+        MessageId::MarkdownTableLabel => "表格",
+        MessageId::ToolOutputStdoutCapturedSummary => "stdout 已捕获",
+        MessageId::ToolOutputStderrCapturedSummary => "stderr 已捕获",
+        MessageId::ToolSummaryExit => "退出码 {exit}",
+        MessageId::ToolSummaryBlocked => "tool 请求被 shell broker guard 阻止",
+        MessageId::ToolSummaryTimedOut => "tool 请求超时",
+        MessageId::ToolSummaryFailed => "tool 请求失败",
+        _ => return None,
+    })
+}
