+++ src/cosh-ng/crates/cosh-shell/src/i18n/zh/config.rs
@@ -1,0 +1,31 @@
+use super::MessageId;
+
+pub(super) fn message(id: MessageId) -> Option<&'static str> {
+    Some(match id {
+        MessageId::ConfigInvalidLanguageBody => "无效语言: {language}",
+        MessageId::ConfigSupportedLanguagesFooter => "支持: auto, en-US, zh-CN。",
+        MessageId::ConfigUnknownKeyBody => "未知配置项: {key}",
+        MessageId::ConfigHomeMissingBody => "HOME 未设置，无法持久化配置。",
+        MessageId::ConfigHomeMissingFooter => "设置 HOME 或手动编辑配置。",
+        MessageId::ConfigUnchangedTitle => "配置未变更",
+        MessageId::ConfigNoFileChangedBody => "未修改配置文件。",
+        MessageId::ConfigSavedTitle => "配置已保存",
+        MessageId::ConfigSavedValueLine => "已保存 ui.{setting} = \"{value}\"。",
+        MessageId::ConfigCurrentSessionLanguageLine => "当前会话语言: {language}。",
+        MessageId::ConfigSavedFooter => "保存的设置会在下次启动时生效。",
+        MessageId::ConfigSaveFailedTitle => "配置保存失败",
+        MessageId::ConfigSaveFailedBody => "配置保存失败: {error}",
+        MessageId::ConfigSavePromptTitle => "保存配置？",
+        MessageId::ConfigFileLine => "文件: {path}",
+        MessageId::ConfigPendingChangeLine => "ui.{setting}: {before} -> {after}",
+        MessageId::ConfigSaveButton => "保存",
+        MessageId::ConfigCancelButton => "取消",
+        MessageId::ConfigApplyKeysFooter => "按键: Left/Right 选择 | Enter 应用 | Esc 取消",
+        MessageId::ConfigLanguageTitle => "语言",
+        MessageId::ConfigLanguageAutoLine => "auto    跟随 LC_ALL/LC_MESSAGES/LANG",
+        MessageId::ConfigLanguageEnLine => "en-US   英语",
+        MessageId::ConfigLanguageZhLine => "zh-CN   简体中文",
+        MessageId::ConfigLanguageKeysFooter => "按键: Left/Right 选择 | Enter 确认 | Esc 取消",
+        _ => return None,
+    })
+}
